+++ demos/storybook/stories/score-card/with-score-badge.stories.ts
@@ -10,7 +10,12 @@ export const withScoreBadge = (): any => ({
     styles: [
         `${withCustomHeaderStyles}
         .sb-score-card-content mat-list-item {
-            height: 36px!important;
+            height: 2.25rem!important;
+        }
+        ::ng-deep .sb-score-card-content.mat-list-base .mat-list-item .mat-list-icon {
+            height: 1.5rem;
+            width: 1.5rem;
+            font-size: 1.5rem;
         }
         .sb-score-card-content mat-icon {
             margin-right: 32px;
@@ -52,7 +57,7 @@ export const withScoreBadge = (): any => ({
             <pxb-hero pxb-badge [label]="'Grade'" [value]="'98'" [units]="'/100'" [iconSize]="72">
                 <i pxb-primary [style.color]="colors.green[500]" class="pxb-grade_a"></i>
             </pxb-hero>
-            <pxb-info-list-item pxb-action-row chevron="true" hidePadding="true" dense="true"(click)="actionRowClick()">
+            <pxb-info-list-item pxb-action-row chevron="true" hidePadding="true" dense="true" (click)="actionRowClick()">
                 <div pxb-title>View Location</div>
             </pxb-info-list-item>
         </pxb-score-card>