+++ demos/storybook/stories/user-menu/within-a-toolbar.stories.ts
@@ -5,14 +5,17 @@ import { invertRTL } from '../../src/utils';
 export const withinToolbar = (): any => ({
     template: `
         <div>
-            <mat-toolbar style="padding: 0px 16px">
-                <h2>Toolbar Title</h2>
+            <ng-template #avatar>
+                    <mat-icon>person</mat-icon>
+            </ng-template>
+            <mat-toolbar style="padding: 0 16px; height: 4rem;">
+                <div class="mat-title">Toolbar Title</div>
                 <pxb-spacer></pxb-spacer>
                 <pxb-user-menu menuTitle="Jane Doe" menuSubtitle="Account Manager" [(open)]="open">
-                    <mat-icon pxb-avatar>person</mat-icon>
-                    <mat-icon pxb-menu-avatar>person</mat-icon>
+                    <template pxb-avatar [ngTemplateOutlet]="avatar"></template>
+                    <template pxb-menu-avatar [ngTemplateOutlet]="avatar"></template>
                     <div pxb-menu-body>
-                         <mat-nav-list [style.paddingTop.px]="0">
+                         <mat-nav-list [style.paddingTop.px]="0" [style.width.rem]="12">
                             <pxb-info-list-item *ngFor="let item of items" [dense]="true" 
                                 (click)="open=false; item.onSelect();">
                                 <mat-icon pxb-icon [style.transform]="invertRTL()">{{item.icon}}</mat-icon>
@@ -27,7 +30,7 @@ export const withinToolbar = (): any => ({
                     </div>
                 </pxb-user-menu> 
             </mat-toolbar>
-            <div style="font-size: 16px; padding: 16px; height: 100px;">Body Content Goes Here</div>
+            <div style="padding: 16px; height: 100px;" class="mat-body-1">Body Content Goes Here</div>
         </div>
     `,
     props: {